Split large CLI test module by maintenance domain
