+++ meshtastic/tests/test_main.py
@@ -9,8 +9,6 @@ import logging
 import platform
 import re
 import sys
-import threading
-import time
 from pathlib import Path
 from types import SimpleNamespace
 from typing import Any, Callable, cast
@@ -21,7 +19,6 @@ import yaml
 
 import meshtastic.__main__ as main_module
 from meshtastic import mt_config
-from meshtastic.region_presets import RegionPresetInfo
 from meshtastic.__main__ import (
     _create_power_meter,
     _normalize_pref_name,
@@ -47,11 +44,10 @@ from ..node import Node
 # from ..radioconfig_pb2 import UserPreferences
 # import meshtastic.config_pb2
 from ..ota import OTAError, OTATransportError
-from ..protobuf import config_pb2, localonly_pb2, mesh_pb2
+from ..protobuf import config_pb2, localonly_pb2
 from ..protobuf.channel_pb2 import Channel  # pylint: disable=E0611
 from ..serial_interface import SerialInterface
 from ..tcp_interface import TCPInterface
-from ..util import Acknowledgment
 
 # from ..remote_hardware import onGPIOreceive
 # from ..config_pb2 import Config
@@ -66,6 +62,17 @@ def _get_config_field(config: Any, dotted_path: str) -> Any:
     for part in dotted_path.split("."):
         obj = getattr(obj, part)
     return obj
+
+
+def _patch_fast_monotonic(monkeypatch: pytest.MonkeyPatch) -> None:
+    """Advance monotonic time quickly in reconnect/config verification tests."""
+    value = [0.0]
+
+    def _fast() -> float:
+        value[0] += 100.0
+        return value[0]
+
+    monkeypatch.setattr(main_module.time, "monotonic", _fast)
 
 
 def _mock_sendText_helper(
@@ -5772,90 +5779,10 @@ def test_printConfig_skips_non_message_sections(
     assert err == ""
 
 
-def _patch_fast_monotonic(monkeypatch: pytest.MonkeyPatch) -> None:
-    _val = [0.0]
-
-    def _fast():
-        _val[0] += 100.0
-        return _val[0]
-
-    monkeypatch.setattr(main_module.time, "monotonic", _fast)
 
 
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_main_configure_phase3_verified_with_matching_config_values(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    config_path = tmp_path / "phase3_verified.yaml"
-    config_path.write_text(
-        yaml.safe_dump({"config": {"power": {"ls_secs": 222}}}),
-        encoding="utf-8",
-    )
-    target_local = localonly_pb2.LocalConfig()
-    iface, target_node = _build_configure_interface(
-        target_local, localonly_pb2.LocalModuleConfig()
-    )
-    iface.isConnected = threading.Event()
-    iface.isConnected.set()
-    iface.waitForConfig = MagicMock()
-    target_node.requestConfig = MagicMock(
-        side_effect=lambda field_desc: (
-            setattr(target_local.power, "ls_secs", 222)
-            if getattr(field_desc, "name", "") == "power"
-            else None
-        )
-    )
-    _patch_fast_monotonic(monkeypatch)
-    _run_main_configure_file(config_path, iface, monkeypatch)
-    out, _ = capsys.readouterr()
-    assert "All settings verified" in out
 
 
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_main_configure_phase1_direct_write_order(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    config_path = tmp_path / "phase1_order.yaml"
-    config_path.write_text(
-        yaml.safe_dump(
-            {
-                "owner": "OrderTest",
-                "owner_short": "OT",
-                "location": {"lat": 1.0, "lon": 2.0, "alt": 3.0},
-                "canned_messages": "A|B|C",
-                "ringtone": "24:d=16,o=5,b=100:c",
-                "channel_url": "https://meshtastic.org/e/#CgYSAQABAA",
-            }
-        ),
-        encoding="utf-8",
-    )
-    iface, target_node = _build_configure_interface()
-    _patch_fast_monotonic(monkeypatch)
-    _run_main_configure_file(config_path, iface, monkeypatch)
-
-    phase1_methods = (
-        "setOwner",
-        "setFixedPosition",
-        "set_canned_message",
-        "set_ringtone",
-        "setURL",
-    )
-    method_names = [c[0] for c in target_node.method_calls]
-    relevant = [m for m in method_names if m in phase1_methods]
-    expected = [
-        "setOwner",
-        "setOwner",
-        "setFixedPosition",
-        "set_canned_message",
-        "set_ringtone",
-        "setURL",
-    ]
-    assert relevant == expected
 
 
 @pytest.mark.unit
@@ -5883,110 +5810,10 @@ def test_main_configure_owner_values_use_normalized_strings(
     ]
 
 
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_main_configure_channel_url_is_terminal_phase1_write(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    config_path = tmp_path / "terminal_seturl.yaml"
-    config_path.write_text(
-        yaml.safe_dump(
-            {
-                "owner": "TerminalTest",
-                "location": {"lat": 10.0, "lon": 20.0, "alt": 30.0},
-                "channel_url": "https://meshtastic.org/e/#CgYSAQABAA",
-            }
-        ),
-        encoding="utf-8",
-    )
-    iface, target_node = _build_configure_interface()
-    _patch_fast_monotonic(monkeypatch)
-    _run_main_configure_file(config_path, iface, monkeypatch)
-
-    method_names = [c[0] for c in target_node.method_calls]
-    seturl_indices = [i for i, m in enumerate(method_names) if m == "setURL"]
-    assert len(seturl_indices) == 1
-    seturl_idx = seturl_indices[0]
-    after_seturl = method_names[seturl_idx + 1 :]
-    for method in (
-        "setFixedPosition",
-        "set_canned_message",
-        "set_ringtone",
-        "setOwner",
-    ):
-        assert method not in after_seturl
 
 
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_main_configure_seturl_unstable_aborts_before_phase2(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    config_path = tmp_path / "unstable_seturl.yaml"
-    config_path.write_text(
-        yaml.safe_dump(
-            {
-                "channel_url": "https://meshtastic.org/e/#CgYSAQABAA",
-                "config": {"power": {"ls_secs": 222}},
-            }
-        ),
-        encoding="utf-8",
-    )
-    target_local = localonly_pb2.LocalConfig()
-    iface, target_node = _build_configure_interface(
-        target_local, localonly_pb2.LocalModuleConfig()
-    )
-    iface.isConnected = threading.Event()
-    iface.isConnected.set()
-    iface.waitForConfig = MagicMock()
-    monkeypatch.setattr(
-        "meshtastic.__main__._post_seturl_stability_check",
-        lambda *a, **k: False,
-    )
-    _patch_fast_monotonic(monkeypatch)
-    with pytest.raises(SystemExit):
-        _run_main_configure_file(config_path, iface, monkeypatch)
-
-    target_node.beginSettingsTransaction.assert_not_called()
-    _, err = capsys.readouterr()
-    assert "transport did not stabilize" in err
 
 
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_main_configure_seturl_stable_proceeds_to_phase2(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    config_path = tmp_path / "stable_seturl.yaml"
-    config_path.write_text(
-        yaml.safe_dump(
-            {
-                "channel_url": "https://meshtastic.org/e/#CgYSAQABAA",
-                "config": {"power": {"ls_secs": 222}},
-            }
-        ),
-        encoding="utf-8",
-    )
-    target_local = localonly_pb2.LocalConfig()
-    iface, target_node = _build_configure_interface(
-        target_local, localonly_pb2.LocalModuleConfig()
-    )
-    iface.isConnected = threading.Event()
-    iface.isConnected.set()
-    iface.waitForConfig = MagicMock()
-    monkeypatch.setattr(
-        "meshtastic.__main__._post_seturl_stability_check",
-        lambda *a, **k: True,
-    )
-    _patch_fast_monotonic(monkeypatch)
-    _run_main_configure_file(config_path, iface, monkeypatch)
-
-    target_node.beginSettingsTransaction.assert_called_once()
-    target_node.commitSettingsTransaction.assert_called_once()
 
 
 @pytest.mark.unit
@@ -6045,489 +5872,34 @@ def test_main_configure_rejects_owner_short_aliases(
     target_node.setOwner.assert_not_called()
 
 
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_main_configure_phase3_no_reconnect_needed(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    config_path = tmp_path / "phase3_no_reboot.yaml"
-    config_path.write_text(
-        yaml.safe_dump({"owner": "TestUser"}),
-        encoding="utf-8",
-    )
-    iface, target_node = _build_configure_interface()
-    _run_main_configure_file(config_path, iface, monkeypatch)
-    out, _ = capsys.readouterr()
-    assert "no reboot expected" in out
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_main_configure_channel_url_only_reports_possible_reconnect(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    config_path = tmp_path / "phase1_channel_url_only.yaml"
-    config_path.write_text(
-        yaml.safe_dump({"channel_url": "https://meshtastic.org/e/#CgcSAQE6AggN"}),
-        encoding="utf-8",
-    )
-    iface, target_node = _build_configure_interface()
-    _run_main_configure_file(config_path, iface, monkeypatch)
-    out, _ = capsys.readouterr()
-    assert "Phase 1: Applying direct configuration" in out
-    assert (
-        "Configuration applied. Channel URL updates may still trigger reconnect/reboot."
-        in out
-    )
-    assert "Configuration applied (no reboot expected)." not in out
-    target_node.setURL.assert_called_once()
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_main_configure_channel_url_skip_when_already_matching(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    config_path = tmp_path / "phase1_channel_url_skip.yaml"
-    config_path.write_text(
-        yaml.safe_dump({"channel_url": "https://meshtastic.org/e/#CgcSAQE6AggN"}),
-        encoding="utf-8",
-    )
-    iface, target_node = _build_configure_interface()
-    monkeypatch.setattr(
-        "meshtastic.__main__._channel_url_matches_current_device_state",
-        lambda *a, **k: True,
-    )
-    _run_main_configure_file(config_path, iface, monkeypatch)
-    out, _ = capsys.readouterr()
-    assert "Channel url already matches device state; skipping apply." in out
-    assert "Configuration applied (no reboot expected)." in out
-    target_node.setURL.assert_not_called()
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_main_configure_phase3_channel_url_verified(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    from ..protobuf import apponly_pb2, channel_pb2
-
-    config_path = tmp_path / "phase3_channel_url.yaml"
-    channel_settings = channel_pb2.ChannelSettings()
-    channel_settings.psk = b"\x01"
-    channel_settings.name = "test"
-    cs = apponly_pb2.ChannelSet()
-    cs.settings.add().CopyFrom(channel_settings)
-    cs.lora_config.region = config_pb2.Config.LoRaConfig.RegionCode.Value("US")
-    cs.lora_config.hop_limit = 3
-    raw = cs.SerializeToString()
-    b64 = base64.b64encode(raw, altchars=b"-_").decode().rstrip("=")
-    test_url = f"https://meshtastic.org/e/#{b64}"
-    config_path.write_text(
-        yaml.safe_dump(
-            {
-                "channel_url": test_url,
-                "config": {"power": {"ls_secs": 222}},
-            }
-        ),
-        encoding="utf-8",
-    )
-    target_local = localonly_pb2.LocalConfig()
-    target_local.lora.region = config_pb2.Config.LoRaConfig.RegionCode.Value("US")
-    target_local.lora.hop_limit = 3
-    iface, target_node = _build_configure_interface(
-        target_local, localonly_pb2.LocalModuleConfig()
-    )
-    primary_channel = channel_pb2.Channel()
-    primary_channel.role = channel_pb2.Channel.Role.PRIMARY
-    primary_channel.settings.CopyFrom(channel_settings)
-
-    def _request_channels_side_effect(*_args: object) -> None:
-        target_node.channels = [primary_channel]
-
-    target_node.channels = [primary_channel]
-    iface.isConnected = threading.Event()
-    iface.isConnected.set()
-    iface.waitForConfig = MagicMock()
-    target_node.requestConfig = MagicMock(
-        side_effect=lambda field_desc: (
-            setattr(target_local.power, "ls_secs", 222)
-            if getattr(field_desc, "name", "") == "power"
-            else None
-        )
-    )
-    target_node.requestChannels = MagicMock(side_effect=_request_channels_side_effect)
-    monkeypatch.setattr(
-        "meshtastic.__main__._verify_channel_url_against_state",
-        lambda *a, **k: False,
-    )
-    monkeypatch.setattr(
-        "meshtastic.__main__._post_seturl_stability_check",
-        lambda *a, **k: True,
-    )
-    _patch_fast_monotonic(monkeypatch)
-    _run_main_configure_file(config_path, iface, monkeypatch)
-    out, _ = capsys.readouterr()
-    assert "Could not fully verify" in out
-
-
-@pytest.mark.unit
-def test_post_seturl_stability_check_triggers_reconnect_when_disconnected(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    event = threading.Event()
-    iface = SimpleNamespace(
-        isConnected=event,
-        waitForConfig=MagicMock(),
-    )
-    iface.connect = MagicMock(side_effect=event.set)
-    monkeypatch.setattr("time.sleep", lambda _: None)
-
-    assert (
-        main_module._post_seturl_stability_check(cast(Any, iface), timeout=2.0) is True
-    )
-    iface.connect.assert_called()
-    iface.waitForConfig.assert_called_once()
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-@pytest.mark.parametrize(
-    ("reset_flag", "expected_full"),
-    [("--factory-reset", False), ("--factory-reset-device", True)],
-)
-def test_main_factory_reset_local_accepts_ack_before_close(
-    capsys: pytest.CaptureFixture[str],
-    reset_flag: str,
-    expected_full: bool,
-) -> None:
-    """Both local reset variants should accept their correlated ACK before closing."""
-    sys.argv = ["", reset_flag]
-    mt_config.args = sys.argv  # type: ignore[assignment]
-    iface = MagicMock(autospec=SerialInterface)
-    iface.__enter__ = MagicMock(return_value=iface)
-    iface.__exit__ = MagicMock(return_value=None)
-    iface.isConnected = threading.Event()
-    iface.isConnected.set()
-    iface._acknowledgment = Acknowledgment()
-    iface._acknowledgment.receivedAck = True  # stale state must be cleared
-    iface._timeout.expireTimeout = 1.0
-    reset_node = iface.getNode.return_value
-    reset_node.iface = iface
-
-    def _factory_reset(*, full: bool) -> SimpleNamespace:
-        assert full is expected_full
-        assert iface._acknowledgment.receivedAck is False
-        iface._acknowledgment.receivedImplAck = True
-        return SimpleNamespace(id=123)
-
-    reset_node.factoryReset.side_effect = _factory_reset
-
-    with patch("meshtastic.serial_interface.SerialInterface", return_value=iface):
-        main()
-
-    out, err = capsys.readouterr()
-    reset_node.factoryReset.assert_called_once_with(full=expected_full)
-    iface.waitForAckNak.assert_not_called()
-    assert iface._acknowledgment.receivedImplAck is False
-    assert "Waiting for factory reset acknowledgment or reboot disconnect" in out
-    assert err == ""
-
-
-@pytest.mark.unit
-def test_local_factory_reset_accepts_transient_reboot_disconnect() -> None:
-    """A 2.7-style dropped ACK is accepted only after reboot disconnect is observed."""
-    connected = threading.Event()
-    connected.set()
-    iface = SimpleNamespace(
-        isConnected=connected,
-        _acknowledgment=Acknowledgment(),
-        _timeout=SimpleNamespace(expireTimeout=1.0),
-        _retire_wait_request=MagicMock(),
-        MeshInterfaceError=RuntimeError,
-    )
-    reset_node = SimpleNamespace(iface=iface)
-
-    def _factory_reset(*, full: bool) -> SimpleNamespace:
-        assert full is False
-        request = SimpleNamespace(id=456)
-
-        def _disconnect_after_send() -> None:
-            time.sleep(0.01)
-            # Model a fast reconnect: the pubsub edge must remain observable even
-            # though isConnected is already set again by the time the wait polls.
-            main_module.pub.sendMessage("meshtastic.connection.lost", interface=iface)
-            connected.set()
-
-        threading.Thread(target=_disconnect_after_send, daemon=True).start()
-        return request
-
-    reset_node.factoryReset = MagicMock(side_effect=_factory_reset)
-
-    request = main_module._send_local_factory_reset_and_wait(
-        reset_node,
-        full=False,
-        timeout=0.5,
-    )
-
-    assert request is not None
-    assert request.id == 456
-    iface._retire_wait_request.assert_called_once_with("receivedNak", request_id=456)
-
-
-@pytest.mark.unit
-def test_local_factory_reset_accepts_implicit_ack_with_legacy_completion_flag() -> None:
-    """A valid implicit ACK must win over the legacy receivedNak completion latch."""
-    connected = threading.Event()
-    connected.set()
-    acknowledgment = Acknowledgment()
-    iface = SimpleNamespace(
-        isConnected=connected,
-        _acknowledgment=acknowledgment,
-        _retire_wait_request=MagicMock(),
-        _raise_wait_error_if_present=MagicMock(),
-        MeshInterfaceError=RuntimeError,
-    )
-
-    def _factory_reset(*, full: bool) -> SimpleNamespace:
-        assert full is False
-        acknowledgment.receivedImplAck = True
-        acknowledgment.receivedNak = True
-        return SimpleNamespace(id=457)
-
-    reset_node = SimpleNamespace(
-        iface=iface,
-        factoryReset=MagicMock(side_effect=_factory_reset),
-    )
-
-    request = main_module._send_local_factory_reset_and_wait(
-        reset_node,
-        full=False,
-        timeout=0.5,
-    )
-
-    assert request is not None
-    assert request.id == 457
-    iface._raise_wait_error_if_present.assert_called()
-
-
-@pytest.mark.unit
-def test_local_factory_reset_accepts_tcp_socket_generation_change() -> None:
-    """A TCP reconnect is observable even when isConnected never clears."""
-    connected = threading.Event()
-    connected.set()
-    original_socket = object()
-    iface = SimpleNamespace(
-        isConnected=connected,
-        socket=original_socket,
-        stream=None,
-        _acknowledgment=Acknowledgment(),
-        _timeout=SimpleNamespace(expireTimeout=0.01),
-        _retire_wait_request=MagicMock(),
-        MeshInterfaceError=RuntimeError,
-    )
-    reset_node = SimpleNamespace(iface=iface)
-
-    def _factory_reset(*, full: bool) -> SimpleNamespace:
-        assert full is False
-
-        def _replace_socket() -> None:
-            time.sleep(0.05)
-            iface.socket = object()
-
-        threading.Thread(target=_replace_socket, daemon=True).start()
-        return SimpleNamespace(id=458)
-
-    reset_node.factoryReset = MagicMock(side_effect=_factory_reset)
-
-    started = time.monotonic()
-    request = main_module._send_local_factory_reset_and_wait(
-        reset_node,
-        full=False,
-    )
-
-    assert request is not None
-    assert request.id == 458
-    assert time.monotonic() - started >= 0.04
-    # The interface's ordinary 10 ms timeout must not truncate the reset's
-    # firmware-defined seven-second reboot window.
-    assert iface._timeout.expireTimeout == 0.01
-
-
-@pytest.mark.unit
-def test_local_factory_reset_ignores_socket_change_before_send_returns() -> None:
-    """A transport change before a concrete request returns is not acceptance."""
-    connected = threading.Event()
-    connected.set()
-    iface = SimpleNamespace(
-        isConnected=connected,
-        socket=object(),
-        stream=None,
-        _acknowledgment=Acknowledgment(),
-        _retire_wait_request=MagicMock(),
-        MeshInterfaceError=RuntimeError,
-    )
-
-    def _factory_reset(*, full: bool) -> SimpleNamespace:
-        assert full is False
-        iface.socket = object()
-        return SimpleNamespace(id=459)
-
-    reset_node = SimpleNamespace(
-        iface=iface,
-        factoryReset=MagicMock(side_effect=_factory_reset),
-    )
-
-    with pytest.raises(
-        RuntimeError,
-        match="Timed out waiting for a factory reset acknowledgment or reboot disconnect",
-    ):
-        main_module._send_local_factory_reset_and_wait(
-            reset_node,
-            full=False,
-            timeout=0.01,
-        )
-
-
-@pytest.mark.unit
-def test_local_factory_reset_ignores_pre_send_disconnect_event() -> None:
-    """A stale disconnect published before the request is queued cannot prove acceptance."""
-    connected = threading.Event()
-    connected.set()
-    iface = SimpleNamespace(
-        isConnected=connected,
-        _acknowledgment=Acknowledgment(),
-        _timeout=SimpleNamespace(expireTimeout=0.01),
-        _retire_wait_request=MagicMock(),
-        MeshInterfaceError=RuntimeError,
-    )
-
-    def _factory_reset(*, full: bool) -> SimpleNamespace:
-        assert full is False
-        main_module.pub.sendMessage("meshtastic.connection.lost", interface=iface)
-        return SimpleNamespace(id=654)
-
-    reset_node = SimpleNamespace(
-        iface=iface,
-        factoryReset=MagicMock(side_effect=_factory_reset),
-    )
-
-    with pytest.raises(
-        RuntimeError,
-        match="Timed out waiting for a factory reset acknowledgment or reboot disconnect",
-    ):
-        main_module._send_local_factory_reset_and_wait(
-            reset_node,
-            full=False,
-            timeout=0.01,
-        )
-
-    iface._retire_wait_request.assert_called_once_with("receivedNak", request_id=654)
-
-
-@pytest.mark.unit
-def test_local_factory_reset_times_out_without_ack_or_disconnect() -> None:
-    """A queued reset without ACK, NAK, or reboot remains a hard failure."""
-    connected = threading.Event()
-    connected.set()
-    iface = SimpleNamespace(
-        isConnected=connected,
-        _acknowledgment=Acknowledgment(),
-        _timeout=SimpleNamespace(expireTimeout=0.01),
-        _retire_wait_request=MagicMock(),
-        MeshInterfaceError=RuntimeError,
-    )
-    reset_node = SimpleNamespace(
-        iface=iface,
-        factoryReset=MagicMock(return_value=SimpleNamespace(id=789)),
-    )
-
-    with pytest.raises(
-        RuntimeError,
-        match="Timed out waiting for a factory reset acknowledgment or reboot disconnect",
-    ):
-        main_module._send_local_factory_reset_and_wait(
-            reset_node,
-            full=False,
-            timeout=0.01,
-        )
-
-    iface._retire_wait_request.assert_called_once_with("receivedNak", request_id=789)
-
-
-@pytest.mark.unit
-def test_local_factory_reset_nak_remains_failure() -> None:
-    """A request-scoped routing error must win over any transport signal."""
-    connected = threading.Event()
-    connected.set()
-    iface = SimpleNamespace(
-        isConnected=connected,
-        socket=object(),
-        stream=None,
-        _acknowledgment=Acknowledgment(),
-        _wait_for_request_ack=MagicMock(return_value=True),
-        _raise_wait_error_if_present=MagicMock(
-            side_effect=RuntimeError("Routing error on response: NOT_AUTHORIZED")
-        ),
-        _retire_wait_request=MagicMock(),
-        MeshInterfaceError=RuntimeError,
-    )
-    reset_node = SimpleNamespace(
-        iface=iface,
-        factoryReset=MagicMock(return_value=SimpleNamespace(id=987)),
-    )
-
-    with pytest.raises(RuntimeError, match="Routing error on response: NOT_AUTHORIZED"):
-        main_module._send_local_factory_reset_and_wait(
-            reset_node,
-            full=False,
-            timeout=0.5,
-        )
-
-    iface._wait_for_request_ack.assert_called_once()
-    iface._raise_wait_error_if_present.assert_called_once_with(
-        "receivedNak", request_id=987
-    )
-    iface._retire_wait_request.assert_called_once_with("receivedNak", request_id=987)
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_main_factory_reset_skips_ack_wait_when_send_is_disabled() -> None:
-    """A noProto reset returning None must not enter an impossible ACK wait."""
-    sys.argv = ["", "--factory-reset"]
-    mt_config.args = sys.argv  # type: ignore[assignment]
-    iface = MagicMock(autospec=SerialInterface)
-    iface.__enter__ = MagicMock(return_value=iface)
-    iface.__exit__ = MagicMock(return_value=None)
-    reset_node = iface.getNode.return_value
-    reset_node.iface = iface
-    reset_node.factoryReset.return_value = None
-
-    with patch("meshtastic.serial_interface.SerialInterface", return_value=iface):
-        main()
-
-    iface.waitForAckNak.assert_not_called()
-
-
-@pytest.mark.unit
-def test_post_factory_reset_ready_probe_closes_and_probes_reconnect() -> None:
-    iface = cast(Any, object.__new__(SerialInterface))
-    iface.connect = MagicMock()
-    iface.close = MagicMock()
-
-    main_module._post_factory_reset_ready_probe(cast(Any, iface))
-
-    iface.connect.assert_called_once()
-    assert iface.close.call_count >= 2
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 
 @pytest.mark.unit
@@ -6807,552 +6179,3 @@ def test_main_ota_update_file_not_found(
     # Verify no transport was constructed before the file check fired
     tcp_cls.assert_not_called()
     serial_cls.assert_not_called()
-
-
-def _run_region_preset_cli(
-    monkeypatch: pytest.MonkeyPatch,
-    *extra_args: str,
-    region_preset_map: object | None,
-    region_presets: dict[int, RegionPresetInfo] | None = None,
-) -> MagicMock:
-    cli_args = ["meshtastic", "--show-region-presets"]
-    if "--dest" not in extra_args:
-        cli_args.extend(("--dest", MAIN_LOCAL_ADDR))
-    cli_args.extend(extra_args)
-    monkeypatch.setattr(sys, "argv", cli_args)
-    initParser()
-    interface = MagicMock()
-    interface.devPath = ""
-    interface.myInfo = SimpleNamespace(my_node_num=int("25d6e474", 16))
-    interface.regionPresetMap = region_preset_map
-    interface.regionPresets = region_presets or {}
-    main_module.onConnected(interface)
-    return interface
-
-
-def _init_lockdown_cli(monkeypatch: pytest.MonkeyPatch, *cli_args: str) -> None:
-    effective_args = list(cli_args)
-    if "--dest" not in effective_args:
-        effective_args.extend(("--dest", MAIN_LOCAL_ADDR))
-    monkeypatch.setattr(sys, "argv", ["meshtastic", *effective_args])
-    initParser()
-
-
-def _run_lockdown_cli(
-    monkeypatch: pytest.MonkeyPatch,
-    *cli_args: str,
-    status: mesh_pb2.LockdownStatus | None = None,
-) -> tuple[MagicMock, MagicMock, MagicMock]:
-    _init_lockdown_cli(monkeypatch, *cli_args)
-    interface = MagicMock()
-    interface.devPath = ""
-    interface.myInfo = SimpleNamespace(my_node_num=int("25d6e474", 16))
-    build = MagicMock(return_value=object())
-    send = MagicMock(return_value=status)
-    monkeypatch.setattr(main_module, "build_lockdown_auth", build)
-    monkeypatch.setattr(main_module, "send_lockdown_auth", send)
-    main_module.onConnected(interface)
-    return interface, build, send
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_show_region_presets_reports_absent_metadata_and_closes(
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    interface = _run_region_preset_cli(
-        monkeypatch,
-        region_preset_map=None,
-    )
-
-    output = capsys.readouterr().out
-    assert "did not provide usable region/preset compatibility metadata" in output
-    interface.close.assert_called_once_with()
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_lockdown_cli_unlock_from_file(
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    read_file = MagicMock(return_value=b"file-secret")
-    monkeypatch.setattr(main_module, "read_lockdown_passphrase_file", read_file)
-    status = mesh_pb2.LockdownStatus(state=mesh_pb2.LockdownStatus.UNLOCKED)
-
-    interface, build, send = _run_lockdown_cli(
-        monkeypatch,
-        "--lockdown-unlock",
-        "--lockdown-passphrase-file",
-        "/tmp/secret",
-        status=status,
-    )
-
-    read_file.assert_called_once_with("/tmp/secret")
-    build.assert_called_once_with(
-        b"file-secret",
-        boots_remaining=0,
-        valid_until_epoch=0,
-        max_session_seconds=0,
-        lock_now=False,
-        disable=False,
-    )
-    send.assert_called_once_with(
-        interface, build.return_value, timeout=20.0, allow_reboot_without_status=False
-    )
-    assert "Lockdown status: UNLOCKED" in capsys.readouterr().out
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_show_region_presets_reports_empty_or_malformed_metadata(
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    interface = _run_region_preset_cli(
-        monkeypatch,
-        region_preset_map=object(),
-        region_presets={},
-    )
-
-    output = capsys.readouterr().out
-    assert "did not provide usable region/preset compatibility metadata" in output
-    assert "preset choices remain unconstrained" in output
-    interface.close.assert_called_once_with()
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_show_region_presets_rejects_remote_destination(
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    interface = _run_region_preset_cli(
-        monkeypatch,
-        "--dest",
-        "!deadbeef",
-        region_preset_map=object(),
-    )
-
-    output = capsys.readouterr().out
-    assert "available only from the local node" in output
-    interface.close.assert_called_once_with()
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_show_region_presets_formats_known_unknown_and_licensed_values(
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    known = RegionPresetInfo(
-        presets=(config_pb2.Config.LoRaConfig.LONG_FAST, 999),
-        default_preset=config_pb2.Config.LoRaConfig.LONG_FAST,
-        licensed_only=True,
-    )
-    unknown = RegionPresetInfo(
-        presets=(998,),
-        default_preset=997,
-        licensed_only=False,
-    )
-
-    interface = _run_region_preset_cli(
-        monkeypatch,
-        "--dest",
-        MAIN_LOCAL_ADDR,
-        region_preset_map=object(),
-        region_presets={config_pb2.Config.LoRaConfig.US: known, 999: unknown},
-    )
-
-    output = capsys.readouterr().out
-    assert "US: default=LONG_FAST licensed-only; presets=LONG_FAST,PRESET_999" in output
-    assert "REGION_999: default=PRESET_997; presets=PRESET_998" in output
-    interface.close.assert_called_once_with()
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_region_preset_cli_without_flag_does_not_read_capability_state(
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    monkeypatch.setattr(
-        sys,
-        "argv",
-        ["meshtastic", "--dest", MAIN_LOCAL_ADDR],
-    )
-    initParser()
-    interface = MagicMock()
-    interface.devPath = ""
-    interface.myInfo = SimpleNamespace(my_node_num=int("25d6e474", 16))
-
-    main_module.onConnected(interface)
-
-    assert "region/preset" not in capsys.readouterr().out
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_lockdown_cli_rejects_remote_destination(
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    monkeypatch.setattr(
-        sys,
-        "argv",
-        [
-            "meshtastic",
-            "--lockdown-unlock",
-            "--lockdown-passphrase",
-            "secret",
-            "--dest",
-            "!deadbeef",
-        ],
-    )
-    initParser()
-    interface = MagicMock()
-    interface.devPath = ""
-    interface.myInfo = SimpleNamespace(my_node_num=int("25d6e474", 16))
-    with pytest.raises(SystemExit):
-        main_module.onConnected(interface)
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_lockdown_cli_rejects_unacknowledged_command_line_secret(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    monkeypatch.setattr(
-        sys,
-        "argv",
-        [
-            "meshtastic",
-            "--lockdown-unlock",
-            "--lockdown-passphrase",
-            "secret",
-            "--dest",
-            MAIN_LOCAL_ADDR,
-        ],
-    )
-    initParser()
-    interface = MagicMock()
-    interface.devPath = ""
-    interface.myInfo = SimpleNamespace(my_node_num=int("25d6e474", 16))
-    with pytest.raises(SystemExit):
-        main_module.onConnected(interface)
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_lockdown_cli_command_line_secret_formats_unknown_state_and_backoff(
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    validate = MagicMock(return_value=b"secret")
-    monkeypatch.setattr(main_module, "validate_lockdown_passphrase", validate)
-    status = mesh_pb2.LockdownStatus(state=999, backoff_seconds=7)  # type: ignore[arg-type]
-
-    _interface, build, _send = _run_lockdown_cli(
-        monkeypatch,
-        "--lockdown-unlock",
-        "--lockdown-passphrase",
-        "secret",
-        "--insecure-lockdown-passphrase-on-command-line",
-        status=status,
-    )
-
-    validate.assert_called_once_with(b"secret")
-    assert build.call_args.args == (b"secret",)
-    output = capsys.readouterr().out
-    assert "Lockdown status: STATE_999" in output
-    assert "Retry backoff: 7s" in output
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_lockdown_cli_provision_confirmation_abort(
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    monkeypatch.setattr("builtins.input", lambda _prompt: "no")
-    _init_lockdown_cli(monkeypatch, "--lockdown-provision")
-    interface = MagicMock()
-    interface.devPath = ""
-    interface.myInfo = SimpleNamespace(my_node_num=int("25d6e474", 16))
-
-    with pytest.raises(SystemExit):
-        main_module.onConnected(interface)
-
-    assert "Aborted." in capsys.readouterr().err
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_lockdown_cli_provision_rejects_mismatched_interactive_passphrases(
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    monkeypatch.setattr("builtins.input", lambda _prompt: "yes")
-    prompts = iter(["secret", "different"])
-    monkeypatch.setattr(main_module.getpass, "getpass", lambda _prompt: next(prompts))
-    _init_lockdown_cli(monkeypatch, "--lockdown-provision")
-    interface = MagicMock()
-    interface.devPath = ""
-    interface.myInfo = SimpleNamespace(my_node_num=int("25d6e474", 16))
-
-    with pytest.raises(SystemExit):
-        main_module.onConnected(interface)
-
-    assert "Lockdown passphrases do not match." in capsys.readouterr().err
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_lockdown_cli_provision_reports_auth_failure(
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    monkeypatch.setattr("builtins.input", lambda _prompt: "yes")
-    prompts = iter(["secret", "secret"])
-    monkeypatch.setattr(main_module.getpass, "getpass", lambda _prompt: next(prompts))
-    monkeypatch.setattr(
-        main_module, "validate_lockdown_passphrase", lambda value: value
-    )
-    failed = mesh_pb2.LockdownStatus(state=mesh_pb2.LockdownStatus.UNLOCK_FAILED)
-    _init_lockdown_cli(monkeypatch, "--lockdown-provision")
-    interface = MagicMock()
-    interface.devPath = ""
-    interface.myInfo = SimpleNamespace(my_node_num=int("25d6e474", 16))
-    monkeypatch.setattr(
-        main_module, "build_lockdown_auth", MagicMock(return_value=object())
-    )
-    monkeypatch.setattr(
-        main_module, "send_lockdown_auth", MagicMock(return_value=failed)
-    )
-
-    with pytest.raises(SystemExit):
-        main_module.onConnected(interface)
-
-    captured = capsys.readouterr()
-    assert "Lockdown status: UNLOCK_FAILED" in captured.out
-    assert "Lockdown authentication failed." in captured.err
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_lockdown_cli_lock_now_allows_reboot_without_status(
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    _interface, build, send = _run_lockdown_cli(
-        monkeypatch,
-        "--lockdown-lock-now",
-        "--lockdown-yes",
-        status=None,
-    )
-    assert build.call_args.kwargs["lock_now"] is True
-    assert send.call_args.kwargs["allow_reboot_without_status"] is True
-    assert "device may already be rebooting" in capsys.readouterr().out
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_lockdown_cli_disable_sets_disable_flag(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    monkeypatch.setattr(
-        main_module, "read_lockdown_passphrase_file", lambda _path: b"secret"
-    )
-    status = mesh_pb2.LockdownStatus(state=mesh_pb2.LockdownStatus.UNLOCKED)
-    _interface, build, _send = _run_lockdown_cli(
-        monkeypatch,
-        "--lockdown-disable",
-        "--lockdown-yes",
-        "--lockdown-passphrase-file",
-        "/tmp/secret",
-        status=status,
-    )
-    assert build.call_args.kwargs["disable"] is True
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_lockdown_cli_accepts_explicit_local_destination(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    status = mesh_pb2.LockdownStatus(state=mesh_pb2.LockdownStatus.UNLOCKED)
-    interface, _build, send = _run_lockdown_cli(
-        monkeypatch,
-        "--lockdown-lock-now",
-        "--lockdown-yes",
-        "--dest",
-        MAIN_LOCAL_ADDR,
-        status=status,
-    )
-    send.assert_called_once()
-    interface.close.assert_called_once_with()
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-@pytest.mark.parametrize(
-    ("cli_args", "error", "expected"),
-    (
-        (
-            (
-                "--lockdown-unlock",
-                "--lockdown-passphrase",
-                "",
-                "--insecure-lockdown-passphrase-on-command-line",
-            ),
-            ValueError("lockdown passphrase must be 1..32 bytes"),
-            "Invalid lockdown options",
-        ),
-        (
-            (
-                "--lockdown-unlock",
-                "--lockdown-passphrase-file",
-                "/tmp/insecure-secret",
-            ),
-            PermissionError(
-                "/tmp/insecure-secret mode is 0o640; "
-                "lockdown passphrase files must be operator-only (0600)"
-            ),
-            "operator-only (0600)",
-        ),
-    ),
-)
-def test_lockdown_cli_maps_passphrase_errors_to_user_facing_exit(
-    monkeypatch: pytest.MonkeyPatch,
-    cli_args: tuple[str, ...],
-    error: Exception,
-    expected: str,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    _init_lockdown_cli(monkeypatch, *cli_args)
-    interface = MagicMock()
-    interface.devPath = ""
-    interface.myInfo = SimpleNamespace(my_node_num=1)
-    if "--lockdown-passphrase-file" in cli_args:
-        monkeypatch.setattr(
-            main_module, "read_lockdown_passphrase_file", MagicMock(side_effect=error)
-        )
-    else:
-        monkeypatch.setattr(
-            main_module, "validate_lockdown_passphrase", MagicMock(side_effect=error)
-        )
-
-    with pytest.raises(SystemExit):
-        main_module.onConnected(interface)
-
-    assert expected in capsys.readouterr().err
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_lockdown_cli_maps_invalid_limits_to_user_facing_exit(
-    monkeypatch: pytest.MonkeyPatch,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    monkeypatch.setattr(
-        main_module, "read_lockdown_passphrase_file", lambda _path: b"secret"
-    )
-    _init_lockdown_cli(
-        monkeypatch,
-        "--lockdown-unlock",
-        "--lockdown-passphrase-file",
-        "/tmp/secret",
-        "--lockdown-boots",
-        "-1",
-    )
-    interface = MagicMock()
-    interface.devPath = ""
-    interface.myInfo = SimpleNamespace(my_node_num=1)
-
-    with pytest.raises(SystemExit):
-        main_module.onConnected(interface)
-
-    output = capsys.readouterr().err
-    assert "Invalid lockdown options" in output
-    assert "boots_remaining must be between 0 and 255" in output
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-@pytest.mark.parametrize(
-    "error",
-    (
-        TimeoutError("no LockdownStatus received before timeout"),
-        ValueError("lockdown authentication is USB-serial only"),
-        RuntimeError("device did not provide my_info"),
-    ),
-)
-def test_lockdown_cli_maps_send_failures_to_user_facing_exit(
-    monkeypatch: pytest.MonkeyPatch,
-    error: Exception,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    monkeypatch.setattr(
-        main_module, "read_lockdown_passphrase_file", lambda _path: b"secret"
-    )
-    monkeypatch.setattr(main_module, "send_lockdown_auth", MagicMock(side_effect=error))
-    _init_lockdown_cli(
-        monkeypatch,
-        "--lockdown-unlock",
-        "--lockdown-passphrase-file",
-        "/tmp/secret",
-    )
-    interface = MagicMock()
-    interface.devPath = ""
-    interface.myInfo = SimpleNamespace(my_node_num=1)
-
-    with pytest.raises(SystemExit):
-        main_module.onConnected(interface)
-
-    output = capsys.readouterr().err
-    assert "Lockdown command failed" in output
-    assert str(error) in output
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_lockdown_cli_interactive_unlock_uses_single_prompt(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    getpass_mock = MagicMock(return_value="secret")
-    validate_mock = MagicMock(return_value=b"secret")
-    monkeypatch.setattr(main_module.getpass, "getpass", getpass_mock)
-    monkeypatch.setattr(main_module, "validate_lockdown_passphrase", validate_mock)
-    status = mesh_pb2.LockdownStatus(state=mesh_pb2.LockdownStatus.UNLOCKED)
-
-    _interface, build, _send = _run_lockdown_cli(
-        monkeypatch,
-        "--lockdown-unlock",
-        status=status,
-    )
-
-    getpass_mock.assert_called_once_with("Lockdown passphrase: ")
-    validate_mock.assert_called_once_with(b"secret")
-    assert build.call_args.args == (b"secret",)
-
-
-@pytest.mark.unit
-@pytest.mark.usefixtures("reset_mt_config")
-def test_lockdown_cli_without_action_does_not_call_lockdown_helpers(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    _init_lockdown_cli(monkeypatch)
-    interface = MagicMock()
-    interface.devPath = ""
-    interface.myInfo = SimpleNamespace(my_node_num=int("25d6e474", 16))
-    build = MagicMock()
-    send = MagicMock()
-    monkeypatch.setattr(main_module, "build_lockdown_auth", build)
-    monkeypatch.setattr(main_module, "send_lockdown_auth", send)
-
-    main_module.onConnected(interface)
-
-    build.assert_not_called()
-    send.assert_not_called()
-    interface.close.assert_not_called()
